Use `com.vanniktech.maven.publish` for publication
In this commit no new publication is set up, just
different publication setup is used.
Still, there are some changes in the published artifacts:
- added gradle metadata (without it KMP library won't work)
- `-wasmJs` suffix changed to `-wasm-js`
- added info to the pom files

Also, in GitHub actions `publishAllPublicationsToSpaceRepository`
is used. This way on every commit we will publish only to space,
and in the future, publication to maven Central will be a manual step